% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{best.models.cpp}
\alias{best.models.cpp}
\title{Best models from a snpmpd object by cpp or maximum number of models - modification of best.models from GUESSFM by Chris Wallace}
\usage{
best.models.cpp(d, cpp.thr = 0.99, maxmod = NULL)
}
\arguments{
\item{d}{snpmod object}

\item{cpp.thr}{cumulative posterior probability threshold for selecting top models; this is ignored when maxmod is specified}

\item{maxmod}{maximum number of top models to output; NULL by default}
}
\value{
list of two objects: models = data.frame of top SNP models ordered by posterior probability (PP); the PPs are re-scaled to sum to 1; old.cpp = cpp before PPs adjusted to sum to 1
}
\description{
Best models from a snpmpd object by cpp or maximum number of models - modification of best.models from GUESSFM by Chris Wallace
}
