% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\docType{data}
\name{beta}
\alias{beta}
\title{List of vectors of single-SNP effects (GWAS) for two simulated traits}
\format{An object of class \code{list} of length 2.}
\usage{
beta
}
\description{
Simulated data was generated using the mvrnorm function in R and genotype matrix X. The two traits have correlation 0.4 and 
each has two causal variants, of which one, rs61839660 is shared between traits; trait 1 has second 
causal variant rs62626317 and trait 2 has second causal variant rs11594656. The traits are from a sample
of 2000, with no missing measurements.
}
\details{
The two vectors are named vectors that have the SNPs listed in the same order and names given by SNP ids.
}
\keyword{datasets}
