% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{calcAdjPP}
\alias{calcAdjPP}
\title{Calculates trait-adjusted posterior probabilities for all traits at sharing parameter kappa}
\usage{
calcAdjPP(
  qt,
  STR,
  SS,
  tau,
  nsnpspermodel,
  kappa,
  PP,
  beta,
  SSy,
  Sxy,
  xcovo,
  Mx,
  N,
  allVres,
  covY,
  Nqq,
  Nq3,
  Nq4
)
}
\arguments{
\item{qt}{vector of trait names}

\item{STR}{list consisting of vectors of model configurations for each trait}

\item{SS}{list consisting of lists of model configuration SNPs for each trait}

\item{tau}{matrix of adjustment terms}

\item{nsnpspermodel}{list of number of SNPs per model for each model in STR}

\item{kappa}{single value of sharing parameter kappa}

\item{PP}{list consisting of vectors of posterior probabilities for the model configurations for each trait}

\item{beta}{list of joint effect estimates for models in STR; multi.beta output}

\item{SSy}{matrix of trait cross-products}

\item{Sxy}{matrix with each column being the cross-product between SNPs and a trait}

\item{xcovo}{SNP covariance matrix}

\item{Mx}{vector of SNP means}

\item{N}{number of individuals with measurements for all traits}

\item{allVres}{list of variance residuals}

\item{covY}{covariance matrix of traits}

\item{Nqq}{matrix of all pair-wise counts of number of individuals with both traits in a pair measured;}

\item{Nq3}{vector of counts of number of individuals with three traits measured; all triples considered; NULL if M < 4}

\item{Nq4}{vector of counts of number of individuals with four traits measured; all quadruples considered; NULL if M < 5}
}
\value{
list of trait-adjusted posterior probabilities for each trait at sharing parameter kappa
}
\description{
Calculates trait-adjusted posterior probabilities for all traits at sharing parameter kappa
}
\author{
Jenn Asimit
}
