% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{calcCres12}
\alias{calcCres12}
\title{covariance between residuals of a pair of models for  a trait pair}
\usage{
calcCres12(imod1, imod2, T1, T2, beta, SSy, Sxy, xcovo, Mx, Nqq)
}
\arguments{
\item{imod1}{index of model from STR for trait T1, to be called from beta}

\item{imod2}{index of model from STR for trait T2, to be called from beta}

\item{T1}{index of first trait}

\item{T2}{index of second trait}

\item{beta}{list of joint beta estimates for each trait}

\item{SSy}{matrix of trait cross-products}

\item{Sxy}{matrix with each column being the cross-product between SNPs and a trait}

\item{xcovo}{covariance matrix of c("one"=1,X)}

\item{Mx}{vector of SNP means}

\item{Nqq}{has N[i,j] = no. with  both trait i and trait j measured}
}
\description{
covariance between residuals of a pair of models for  a trait pair
}
\author{
Jenn Asimit
}
