% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{calcD12}
\alias{calcD12}
\title{internal function for calcAdjPP for the 2-trait case}
\usage{
calcD12(
  mod1,
  mod2,
  T1,
  T2,
  beta,
  SSy,
  Sxy,
  xcovo,
  Mx,
  Nqq,
  Vres,
  covY,
  nsnpspermodel
)
}
\arguments{
\item{mod1}{vector of model indices for trait 1}

\item{mod2}{vector of model indices for trait 2}

\item{T1}{index of trait 1}

\item{T2}{index of trait 2}

\item{beta}{list of joint beta estimates for each trait}

\item{SSy}{matrix of trait cross-products}

\item{Sxy}{matrix with each column being the cross-product between SNPs and a trait}

\item{xcovo}{covariance matrix of c("one"=1,X)}

\item{Mx}{vector of SNP means}

\item{Nqq}{has N[i,j] = no. with  both trait i and trait j measured}

\item{Vres}{list of variance residuals}

\item{covY}{covariance matrix of traits}

\item{nsnpspermodel}{list of number of SNPs per model for each model in mod1, mod2}
}
\value{
lbf12-lbf1-lbf2
}
\description{
internal function for calcAdjPP for the 2-trait case
}
\author{
Jenn Asimit
}
