% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamexpanded.corX.R
\name{cor.refdata.fn}
\alias{cor.refdata.fn}
\title{Thin genotype correlation matrix for JAM input}
\usage{
cor.refdata.fn(corX, r2 = 0.99)
}
\arguments{
\item{corX}{correlation matrix of genotypes scores}

\item{r2}{r.squared threshold for thinning SNPs}
}
\value{
correlation matrix of SNPs thinned at r2 and with colinearity removed
}
\description{
Thin genotype correlation matrix for JAM input
}
\author{
Jenn Asimit
}
