% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLASHFMwithFINEMAP.R
\name{credset}
\alias{credset}
\title{Construct a credible set for a trait}
\usage{
credset(modPP, cred = 0.99)
}
\arguments{
\item{modPP}{named vector of model PPs}

\item{cred}{probability for credible set; default is 0.99}
}
\value{
cs vector of SNPs belonging to the credible set
}
\description{
Construct a credible set for a trait
}
\author{
Jenn Asimit
}
