% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs.R
\docType{data}
\name{cvs}
\alias{cvs}
\title{Vector of rsids for causal variants of two simulated traits}
\format{An object of class \code{character} of length 3.}
\usage{
cvs
}
\description{
This is the vector c(c1,c2,c3) = c("rs61839660","rs62626317","rs11594656") for the provided simulation example of two traits 
where one trait has causal variants c1+c2 and the other trait has causal variants c1+c3; the traits have shared causal variant
c1.
}
\keyword{datasets}
