% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{expand.mod}
\alias{expand.mod}
\title{internal function for expanding models by tag SNPs in JAMexpanded.multi}
\usage{
expand.mod(snpmod, taglist)
}
\description{
internal function for expanding models by tag SNPs in JAMexpanded.multi
}
