% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{flashfm}
\alias{flashfm}
\title{Marginal PP for models of a set of traits, sharing information between the traits}
\usage{
flashfm(
  main.input,
  TOdds,
  covY,
  ss.stats,
  cpp = 0.99,
  maxmod = NULL,
  fastapprox = FALSE
)
}
\arguments{
\item{main.input}{List of 3 components: SM=list of snpmod objects for a set of traits; mbeta=list of joint effects for each trait; nsnps= number of SNPs in the region 
This could be obtained from flashfm.input or JAMexpanded.multi.}

\item{TOdds}{Vector of target odds of no sharing to sharing}

\item{covY}{trait covariance matrix}

\item{ss.stats}{output from summaryStats; list of 4 components: Mx = mean of SNPs, xcovo = covariance matrix of SNPs, Sxy = matrix of Sxy values (column traits), ybar=vector of trait means}

\item{cpp}{cumulative posterior probability threshold for selecting top models; this is ignored when maxmod is spe$}

\item{maxmod}{maximum number of top models to output; NULL by default}

\item{fastapprox}{logical that is TRUE when fast approximation is used that does not include unequal sample size adjustments; default is FALSE}
}
\value{
List consisting of PP: marginal PP for models and MPP: marginal PP of SNP inclusion
}
\description{
Marginal PP for models of a set of traits, sharing information between the traits
}
\author{
Jenn Asimit
}
