% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\docType{data}
\name{fm}
\alias{fm}
\title{Output from running FLASHFMwithJAM on the simulated dataset in this package}
\format{
An object of class \code{list} of length 2.
}
\usage{
fm
}
\description{
This is a list with two components:
fm$mpp.pp is a list with 4 components giving the SNP-level results (mpp.pp$PP,mpp.pp$MPP) and SNP group level results (mpp.pp$MPPg, mpp.pp$PPg); and
fm$snpGroups is a list with 2 components giving the SNP groups construced under single-trait (snpGroups[[1]]) and multi-trait fine-mapping (snpGroups[[2]])
}
\keyword{datasets}
