% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMconfig.R
\docType{data}
\name{FMconfig}
\alias{FMconfig}
\title{List of FINEMAP config file contents for two simulated traits}
\format{An object of class \code{list} of length 2.}
\usage{
FMconfig
}
\description{
Simulated data was generated using the mvrnorm function in R and genotype matrix X. The two traits have correlation 0.4 and 
each has two causal variants, of which one, rs61839660 is shared between traits; trait 1 has second 
causal variant rs62626317 and trait 2 has second causal variant rs11594656. The traits are from a sample
of 2000, with no missing measurements.
}
\details{
FINEMAP (Benner, 2016) was run on each trait. The first list data.frame is the config file contents 
for trait 1, and the second list data.frame are those for trait 2.
}
\keyword{datasets}
