% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMtables.R
\name{FMtables}
\alias{FMtables}
\title{Formats flashfm output to table format}
\usage{
FMtables(
  PP,
  SW,
  stepwise = F,
  regname,
  path.input,
  path.output,
  trait.id,
  trait.names = NULL
)
}
\arguments{
\item{PP}{MFM:::MPP.PP.groups.fn output from 'flashfm' in Rdata format.}

\item{SW}{stepwise list of results from Stepwise model. By def. = NULL}

\item{stepwise}{TRUE if list of results from sw is provided. By def. = FALSE}

\item{regname}{Region name for table caption}

\item{path.input}{path where the input files are located.}

\item{path.output}{path to save the output latex table.}

\item{trait.id}{id. number if the traits.}

\item{trait.names}{trait names. If not provided a vector is constructed. 'Trait_1', 'Trait_2', ...}
}
\value{
Table results as txt file.
}
\description{
Formats flashfm output to table format
}
\author{
Nico Hernandez
}
