% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{groupIDs.fn}
\alias{groupIDs.fn}
\title{Find SNP group ids for a set of SNPs}
\usage{
groupIDs.fn(snpgroups, Msnps)
}
\arguments{
\item{snpgroups}{list of snpgroups}

\item{Msnps}{vector of snps}
}
\value{
vector same length of msnps that gives the SNP group containing each SNP, or the SNP id if it does not belong to a group
}
\description{
Find SNP group ids for a set of SNPs
}
\author{
Jenn Asimit
}
