% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.multi.R
\name{groupmulti}
\alias{groupmulti}
\title{Group SNPs; adapted from group.multi of https://github.com/chr1swallace/GUESSFM}
\usage{
groupmulti(SM2, snp.data, is.snpmat, min.mppi = 0.001, r2.minmerge = 0.5)
}
\arguments{
\item{SM2}{snpmod object}

\item{snp.data}{SnpMatrix or SNP covariance matrix with snp ids as column names}

\item{is.snpmat}{logical taking value TRUE when SnpMatrix is provided and FALSE when covariance matrix is given}

\item{min.mppi}{trim snp groups with total MPPI < min.mppi in all
diseases}

\item{r2.minmerge}{merge groups with minimum between-group r2 >
r2.minmerge}
}
\value{
list with three components.

First is a data.frame with each row giving summary statistics for
    each group.

Second is a groups object, each elements ordered according to the rows of the summary

Third is the r2 matrix calculated.
}
\description{
Group SNPs; adapted from group.multi of https://github.com/chr1swallace/GUESSFM
}
