% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{logsum}
\alias{logsum}
\title{logsum}
\usage{
logsum(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
max(x) + log(sum(exp(x - max(x))))
}
\description{
Internal function, logsum (copied from coloc package)
}
\details{
This function calculates the log of the sum of the exponentiated
logs taking out the max, i.e. insuring that the sum is not Inf

ie, you want sum(x), but have x already stored in logs.  log(sum(exp(x))) might fail,
but logsum(x) should work.
}
\examples{
x <- 1:10
log(sum(x))
}
\author{
Claudia Giambartolomei
}
