% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{makemod}
\alias{makemod}
\title{internal function makemod}
\usage{
makemod(snps)
}
\arguments{
\item{snps}{model given by snp ids separated by \code{"\%"}}
}
\description{
internal function makemod
}
\author{
Chris Wallace
}
