% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neffective.R
\name{makeNlist.rel}
\alias{makeNlist.rel}
\title{Sample size information needed for flashfm, when samples are related and have effective sample sizes for each trait}
\usage{
makeNlist.rel(Ne, y = NULL, Nsame = NULL)
}
\arguments{
\item{Ne}{vector or trait effective sample sizes}

\item{y}{if available, matrix of trait measurements or indicators of non-NA trait measurements (all samples, not just unrelated);columns are traits; 
used to get joint sample counts; default is NULL and if not provided an approximation is used based on vector of trait sample sizes}

\item{Nsame}{a single sample size that is the same for all traits}
}
\value{
list of 4 components: N = number of individuals with all traits measured; 
Nqq=matrix of all pair-wise counts of number of individuals with both traits in a pair measured;
Nq3 = vector of counts of number of individuals with three traits measured; all triples considered; NULL if M < 4
Nq4 = vector of counts of number of individuals with four traits measured; all quadruples considered; NULL if M < 5
}
\description{
Sample size information needed for flashfm, when samples are related and have effective sample sizes for each trait
}
\author{
Jenn Asimit
}
