% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{makeplink}
\alias{makeplink}
\title{Convert genotype score data to plink ped-map format}
\usage{
makeplink(Gmat, snpinfo, chr)
}
\arguments{
\item{Gmat}{matrix of genotypes scores: SNPs are columns, individuals are rows; could be dosages, which will be rounded}

\item{snpinfo}{a data.frame with columns: snpid, BP, allele1, allele2; BP is base pair position, allele1 is reference allele}

\item{chr}{chromosome number}
}
\value{
list with two components: ped and map, which are the plink format files
}
\description{
Convert genotype score data to plink ped-map format
}
\author{
Jenn Asimit
}
