% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.multi.R
\name{makeSNPgroups}
\alias{makeSNPgroups}
\title{Make SNP groups using fine-mapping information from all of the traits}
\usage{
makeSNPgroups(main.input, is.snpmat, min.mppi = 0.001, r2.minmerge = 0.5)
}
\arguments{
\item{main.input}{output from flashfm.input function}

\item{is.snpmat}{logical taking value TRUE when genotype matrix is provided and FALSE when covariance matrix is given}

\item{min.mppi}{trim snp groups with total MPPI < min.mppi in all diseases; default 0.001}

\item{r2.minmerge}{merge groups with minimum between-group r2 > r2.minmerge; default 0.5}
}
\description{
Make SNP groups using fine-mapping information from all of the traits
}
