% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{makesnpmod}
\alias{makesnpmod}
\title{converts data.frame of abfs to snpmod object; modified abf2snpmod from GUESSFM}
\usage{
makesnpmod(abf, expected, overdispersion = 1, nsnps = NULL)
}
\description{
converts data.frame of abfs to snpmod object; modified abf2snpmod from GUESSFM
}
