% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{marg.snps.vecs}
\alias{marg.snps.vecs}
\title{internal function marg.snps.vec}
\usage{
marg.snps.vecs(str, pp)
}
\arguments{
\item{str}{models given by snp ids separated by \code{"\%"}}

\item{pp}{posterior probabilities}
}
\description{
internal function marg.snps.vec
}
\author{
Chris Wallace
}
