% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{marginalpp}
\alias{marginalpp}
\title{Marginal PP for models sharing information between traits}
\usage{
marginalpp(
  STR,
  PP,
  mbeta,
  covY,
  SSy,
  Sxy,
  kappa,
  N,
  Nqq,
  nsnps,
  Mx,
  xcovo,
  Nq3,
  Nq4
)
}
\arguments{
\item{STR}{list of models for traits 1, 2, ..., n, each given in
the form of a character vector, with entries
\code{"snp1\%snp2\%snp3"}. The null model is given by
\code{"1"} OR \code{"0"}.  It is assumed that all elements of
ABF, PP and pr below follow this same order.}

\item{PP}{list of posterir probabilities for the models in M}

\item{mbeta}{list of joint beta estimates for each trait}

\item{covY}{trait covariance matrix}

\item{SSy}{matrix of trait cross-products}

\item{Sxy}{matrix with each column being the cross-product between SNPs and a trait}

\item{kappa}{single value or vector of values to consider for the
sharing scale parameter.  the value of kappa=1 must be
included, and if not will be prepended.}

\item{N}{number of individiduals with measurements for all traits}

\item{Nqq}{matrix in which Nqq[i,j] = number of individuals  measured in both trait i and  trait j}

\item{nsnps}{number of snps in region}

\item{Mx}{vector of SNP means}

\item{xcovo}{SNP covariance matrix}

\item{Nq3}{vector of counts of number of individuals with three traits measured; all triples considered; NULL if M < 4}

\item{Nq4}{vector of counts of number of individuals with four traits measured; all quadruples considered; NULL if M < 5}
}
\value{
list of: - single.pp: list of pp for each model in
    \code{STR[[i]]} for trait i - shared.pp: list of pp for each model
    in \code{STR[[i]]} for trait i
}
\description{
Marginal PP for models sharing information between traits
}
