% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mod.fn}
\alias{mod.fn}
\title{internal processing function for JAMexpanded.multi}
\usage{
mod.fn(binout)
}
\arguments{
\item{binout}{named vector of indicators for SNP presence in model; names are snp ids}
}
\value{
snp model represented by snps separated by \code{"\%"}
}
\description{
internal processing function for JAMexpanded.multi
}
\author{
Jenn Asimit
}
