% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\docType{data}
\name{N}
\alias{N}
\title{Vector of samples sizes for two simulated traits}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
N
}
\description{
This is the vector c(2000,2000) for the provided simulation example of two traits having 
no missing measurements from a sample of 2000 individuals. N is required to run JAMexpanded.multi 
and, in the case that FINEMAP results are used, the function flashfm.input.
}
\keyword{datasets}
