% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{PPsummarise}
\alias{PPsummarise}
\title{Summarise PP and MPP results from single-trait fine-mapping and flashfm, by SNP and by SNP group}
\usage{
PPsummarise(fm.multi, snpGroups, minPP = 0.05)
}
\arguments{
\item{fm.multi}{output from flashfm function}

\item{snpGroups}{list of two sets of snp groups; output from makeSNPgroups2}

\item{minPP}{a single value such that output consists of snp/group models where PP > minPP for at least one set of fine-mapping results; default is 0.05}
}
\value{
a list of 4 objects: MPP lists trait-specific PP of SNP inclusion in a model; MPPg lists trait-specific PP of SNP group inclusion in a model;  
		PP lists trait-specific model PP; PPg lists trait-specific model PP in terms of  SNP group
}
\description{
Summarise PP and MPP results from single-trait fine-mapping and flashfm, by SNP and by SNP group
}
\author{
Jenn Asimit
}
