% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prune_maximal}
\alias{prune_maximal}
\title{internal function for ref.data.fn}
\usage{
prune_maximal(genotype_matrix)
}
\arguments{
\item{genotype_matrix}{SNP matrix where SNPS are columns}
}
\description{
internal function for ref.data.fn
}
\author{
Simon Schoenbuchner
}
