% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{refdata.fn}
\alias{refdata.fn}
\title{Thin genotype matrix for JAM input}
\usage{
refdata.fn(X, r2 = 0.99)
}
\arguments{
\item{X}{matrix of genotypes scores: SNPs are columns, individuals are rows; could be dosages}

\item{r2}{r.squared threshold for thinning SNPs}
}
\value{
genotype matrix of SNPs thinned at r2 and with colinearity removed
}
\description{
Thin genotype matrix for JAM input
}
\author{
Jenn Asimit
}
