% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{score2alleles}
\alias{score2alleles}
\title{internal function for makeplink}
\usage{
score2alleles(snp, G, alleles)
}
\arguments{
\item{snp}{column index of snp in genotype matrix}

\item{G}{genotype matrix (SNPs in columns, individuals in rows)}

\item{alleles}{data,frame of two columns where each row gives the alleles for each SNP in G}
}
\description{
internal function for makeplink
}
\author{
Jenn Asimit
}
