% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{SNPprior}
\alias{SNPprior}
\title{SNP prior based on binomial distribution; modification of snpprior from GUESSFM package}
\usage{
SNPprior(
  x = 0:10,
  n,
  expected,
  overdispersion = 1,
  pi0 = NA,
  truncate = NA,
  overdispersion.warning = TRUE
)
}
\description{
SNP prior based on binomial distribution; modification of snpprior from GUESSFM package
}
