% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{summaryStats}
\alias{summaryStats}
\title{Summary statistics needed for flashfm input}
\usage{
summaryStats(Xmat = TRUE, ybar.all, main.input)
}
\arguments{
\item{Xmat}{logical, TRUE if main.input is based on genotype matrix (reference or from sample); FALSE if main.input based on covariance matrix}

\item{ybar.all}{vector of trait means}

\item{main.input}{output from flashfm.input}
}
\value{
list of 4 components: Mx = mean of SNPs, xcovo = covariance matrix of SNPs including column and row of 0s for intercept term, Sxy = matrix of Sxy values (column traits), ybar=vector of trait means
}
\description{
Summary statistics needed for flashfm input
}
\author{
Jenn Asimit
}
