% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{Sxy.hat}
\alias{Sxy.hat}
\title{estimates cross-product of each SNP with one trait}
\usage{
Sxy.hat(beta1, Mx, N, Vx, muY)
}
\arguments{
\item{beta1}{vector of effect estimates from single SNP models}
}
\description{
estimates cross-product of each SNP with one trait
}
\author{
Jenn Asimit
}
