% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamexpanded.corX.R
\name{tagexpand.mod}
\alias{tagexpand.mod}
\title{internal function for expanding models by tag SNPs in JAMexpandedCor.multi}
\usage{
tagexpand.mod(snpmod, taglist)
}
\arguments{
\item{snpmod}{intial snp model with snps separated by \code{"\%"}}

\item{taglist}{list of tag snps for each snp}
}
\description{
internal function for expanding models by tag SNPs in JAMexpandedCor.multi
}
\author{
Jenn Asimit
}
