% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUESSFMfn.R
\docType{class}
\name{tags-class}
\alias{tags-class}
\title{Tags focused class for holding information about sets of SNPs
defined by their mutual LD}
\description{
Tags focused class for holding information about sets of SNPs
defined by their mutual LD
}
\section{Slots}{

\describe{
\item{\code{tags}}{character vector giving tag SNPs, one per SNP in
in \code{.Data}, repeated as necessary}

\item{\code{.Data}}{character vector giving SNPs included in this tags object
new("tags")}
}}

