% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{allC12}
\alias{allC12}
\title{internal function for calcAdjPP for that gives list of covariance matrix of residuals for all trait pairs}
\usage{
allC12(M, nummods, beta, SSy, Sxy, xcovo, Mx, Nqq)
}
\description{
internal function for calcAdjPP for that gives list of covariance matrix of residuals for all trait pairs
}
\author{
Jenn Asimit
}
