% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{calcD12}
\alias{calcD12}
\title{internal function for calcAdjPP for the 2-trait case}
\usage{
calcD12(
  mod1,
  mod2,
  T1,
  T2,
  beta,
  SSy,
  Sxy,
  xcovo,
  Mx,
  Nqq,
  Vres,
  covY,
  nsnpspermodel
)
}
\description{
internal function for calcAdjPP for the 2-trait case
}
\author{
Jenn Asimit
}
