% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{calcDcon}
\alias{calcDcon}
\title{internal function for calcAdjPP that gives constant term for delta}
\usage{
calcDcon(covY, Nqq)
}
\description{
internal function for calcAdjPP that gives constant term for delta
}
\author{
Jenn Asimit
}
