% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{calcQ12}
\alias{calcQ12}
\title{internal function for calcAdjPP for a pair of traits}
\usage{
calcQ12(i, j, T1, T2, SS, tau, nsnpspermodel, kappa)
}
\description{
internal function for calcAdjPP for a pair of traits
}
\author{
Jenn Asimit
}
