% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{expand.mod}
\alias{expand.mod}
\title{internal function for expanding models by tag SNPs in JAMexpanded.multi}
\usage{
expand.mod(snpmod, taglist)
}
\arguments{
\item{snpmod}{intial snp model with snps separated by \code{"\%"}}

\item{taglist}{list of tag snps for each snp}
}
\description{
internal function for expanding models by tag SNPs in JAMexpanded.multi
}
\author{
Jenn Asimit
}
