% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{JAM_PointEstimates}
\alias{JAM_PointEstimates}
\title{internal function for multibeta, modified from JAM_PointEstimates_Package_Simplified of R2BGLiMS package}
\usage{
JAM_PointEstimates(marginal.betas = NULL, X.ref = NULL, n = NULL, ybar)
}
\description{
internal function for multibeta, modified from JAM_PointEstimates_Package_Simplified of R2BGLiMS package
}
