% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{Vres.all}
\alias{Vres.all}
\title{variance of model residuals for trait T1 at all models that have joint effect estimates}
\usage{
Vres.all(Nqq, mbeta, SSy, Sxy)
}
\description{
variance of model residuals for trait T1 at all models that have joint effect estimates
}
\author{
Jenn Asimit
}
