% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm.R
\name{Vres.hat}
\alias{Vres.hat}
\title{variance of model residuals for trait T1 at model index imod}
\usage{
Vres.hat(imod, T1, SSy, Sxy, mbeta, Nqq)
}
\description{
variance of model residuals for trait T1 at model index imod
}
\author{
Jenn Asimit
}
